+++ src/test/ai.rs
@@ -129,19 +129,19 @@ fn _create_minimal_world() -> ((i32, i32), GameState, GameObjects) {
 
     // make tiles near the player walkable
     objects
-        .get_tile_at(p_x as usize, p_y as usize)
+        .get_tile_at(p_x, p_y)
         .replace(Tile::floor(p_x, p_y, innit_env().is_debug_mode));
     objects
-        .get_tile_at((p_x + 1) as usize, p_y as usize)
+        .get_tile_at(p_x + 1, p_y)
         .replace(Tile::floor(p_x + 1, p_y, innit_env().is_debug_mode));
     objects
-        .get_tile_at((p_x - 1) as usize, p_y as usize)
+        .get_tile_at(p_x - 1, p_y)
         .replace(Tile::floor(p_x - 1, p_y, innit_env().is_debug_mode));
     objects
-        .get_tile_at(p_x as usize, (p_y - 1) as usize)
+        .get_tile_at(p_x, p_y - 1)
         .replace(Tile::floor(p_x, p_y - 1, innit_env().is_debug_mode));
     objects
-        .get_tile_at(p_x as usize, (p_y + 1) as usize)
+        .get_tile_at(p_x, p_y + 1)
         .replace(Tile::floor(p_x, p_y + 1, innit_env().is_debug_mode));
 
     let player = Object::new()
